Wrote Python Code
Created a (currently untested) version of the Python code, that will
allow for a button to be pressed, triggering the code to run. This will
not only cause the lights to change over the course of about 30 seconds,
but will also relay the current status of the board to the Web
Interface, which will update accordingly.